+++ fast_food/src/main/java/com/fast_food/entite/StockMatierePremiere.java
@@ -1,5 +1,6 @@
 package com.fast_food.entite;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import jakarta.persistence.Entity;
@@ -22,5 +23,5 @@ public class StockMatierePremiere {
     private List<Recette> recettes;
 
     private String nom;
-    private int quantite;
+    private BigDecimal quantite;
 }
